Move token iterators to module `token`
