+++ pythonx/jieba_vim_rs_core/src/motion/mod.rs
@@ -29,7 +29,6 @@ mod omap_d_ge;
 mod omap_e;
 mod omap_ge;
 mod omap_w;
-mod token_iter;
 mod xmap_b;
 mod xmap_e;
 mod xmap_ge;
